05-Agentes/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
supabase>=2.0.0
rich>=13.0.0
schedule>=1.2.0
loguru>=0.7.0
python-dotenv>=1.0.0
